qi-fix-slow-first-bulk.patch
This patch performs the equivalent init actions to the SDHC slow
unitl first bulk patch in kernel, it allows us to work with large
SDHC cards which exceed Glamo timeout capability at 16MHz for first
access.

+++ src/phase2.c
@@ -81,7 +81,8 @@ void bootloader_second_phase(void)
 		if (this_kernel->partition_index) {
 			unsigned char *p = kernel_dram;
 
-			if (this_kernel->block_read(kernel_dram, 0, 4) < 0) {
+			if ((int)this_kernel->block_read(kernel_dram, 0, 4)
+									  < 0) {
 				puts("Bad partition read\n");
 				this_kernel = &this_board->
 							kernel_source[kernel++];
